Added support for binary properties
There are two ways to get the value, the PHPCR conformant way is to just get the Property object and call getBinary() on it. It will then fetch the binary data from the server and return it. It is not cached anywhere for memory usage reasons.

The second option is to call getBinaryProperty('full/path/to/property') on the ObjectManager instance. This is not part of the official API but it's the most performant way to snatch the image without instanciating a Node and Property object.