Adding Node properties to DoctrineDBAL responses
This doesn't feel like the right way to get this working, but it's
a start.

Each item returned in a DoctrineDBAL query must have the appropriate
props extracted from the 'props' field.  This might be doable in the
SQL query, but I was unsure how properties with multiple values
would be handled.