Temporarily removing the column statements from SQL
This was causing another test to fail.  The logic for adding explicit
column names to the SQL statement is not very solid atm.  Additionally,
this may not even be necessary (although restricting the data pulled
down for each result would be preferred over grabbing all of the data
every time).